Use dynamic buffer instead of static buffer
This removes the requirement to modify the aJson library to get a larger
buffer. The default method allocates a buffer for rendering JSON
dynamically based on output size. A new method has also been added for
applications with less available SRAM (setBufferlessResponses). This
requires no large fixed or dynamic buffer, but increases response
rendering time by close to two orders of magnitude due to inefficiencies
and overhead in the WiFiClient and aJson. With the following PR in
place, this is reduced to an order of magnitude difference (50ms vs
400ms for /api): https://github.com/interactive-matter/aJson/pull/91